+++ doc/release_notes.md
@@ -4,6 +4,7 @@
 
 1.  Fixed bugs and enhanced parsing of connect strings
     (`issue 2 <https://github.com/oracle/rust-oracledb/issues/2>`__).
+1.  Added runnable examples.
 
 
 ## rust-oracledb 26.0.0-beta.1 (August 6, 2026)